requirements.txt
numpy==1.23.4
pybullet==3.2.7
pydantic==2.10.6
requests==2.32.3
open3d==0.19.0
scipy==1.13.1
matplotlib==3.10.1
easydict==1.13
opencv_python_headless==4.8.1.78
supervision==0.17.1
